split shuffle & dispatch
