+++ pkg/sql/colexec/dispatch/dispatch.go
@@ -20,7 +20,6 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/matrixorigin/matrixone/pkg/common/moerr"
-	"github.com/matrixorigin/matrixone/pkg/container/batch"
 	"github.com/matrixorigin/matrixone/pkg/logutil"
 	"github.com/matrixorigin/matrixone/pkg/sql/colexec"
 	"github.com/matrixorigin/matrixone/pkg/vm/process"
@@ -57,7 +56,6 @@ func Prepare(proc *process.Process, arg any) error {
 		} else {
 			ap.prepareLocal()
 		}
-		ap.initShuffle()
 
 	case SendToAnyFunc:
 		if ctr.remoteRegsCnt == 0 {
@@ -94,14 +92,27 @@ func Prepare(proc *process.Process, arg any) error {
 func Call(idx int, proc *process.Process, arg any, isFirst bool, isLast bool) (bool, error) {
 	ap := arg.(*Argument)
 
-	bat := proc.InputBatch()
-	if bat == nil {
-		if ap.FuncId == ShuffleToAllFunc {
-			return sendShuffledBats(ap, proc)
+	if ap.FuncId == ShuffleToAllFunc {
+		bats := proc.InputBatches()
+		if len(bats) == 0 {
+			return true, nil
 		}
-		return true, nil
+		for _, bat := range bats {
+			end, err := shuffleToAllFunc(bat, ap, proc)
+			if err != nil {
+				return false, err
+			}
+			if end {
+				return true, nil
+			}
+		}
+		return false, nil
 	}
 
+	bat := proc.InputBatch()
+	if bat == nil {
+		return true, nil
+	}
 	if bat.Length() == 0 {
 		bat.Clean(proc.Mp())
 		return false, nil
@@ -153,22 +164,4 @@ func (arg *Argument) prepareLocal() {
 	arg.ctr.prepared = true
 	arg.ctr.isRemote = false
 	arg.ctr.remoteReceivers = nil
-}
-
-func (arg *Argument) initShuffle() {
-	if arg.ctr.sels == nil {
-		arg.ctr.sels = make([][]int32, arg.ctr.aliveRegCnt)
-		for i := 0; i < arg.ctr.aliveRegCnt; i++ {
-			arg.ctr.sels[i] = make([]int32, 8192)
-		}
-		arg.ctr.batsCount = 0
-		arg.ctr.shuffledBats = make([]*batch.Batch, arg.ctr.aliveRegCnt)
-	}
-}
-
-func (arg *Argument) getSels() [][]int32 {
-	for i := range arg.ctr.sels {
-		arg.ctr.sels[i] = arg.ctr.sels[i][:0]
-	}
-	return arg.ctr.sels
 }
