+++ pkg/sql/colexec/shuffle/shuffle.go
@@ -17,6 +17,10 @@ package shuffle
 import (
 	"bytes"
 	"github.com/matrixorigin/matrixone/pkg/container/batch"
+	"github.com/matrixorigin/matrixone/pkg/container/types"
+	"github.com/matrixorigin/matrixone/pkg/container/vector"
+	"github.com/matrixorigin/matrixone/pkg/pb/plan"
+	plan2 "github.com/matrixorigin/matrixone/pkg/sql/plan"
 	"github.com/matrixorigin/matrixone/pkg/vm/process"
 )
 
@@ -28,22 +32,335 @@ func Prepare(proc *process.Process, arg any) error {
 	ap := arg.(*Argument)
 	ctr := new(container)
 	ap.ctr = ctr
+	ap.initShuffle()
 	return nil
 }
 
 func Call(idx int, proc *process.Process, arg any, isFirst bool, isLast bool) (bool, error) {
-	//ap := arg.(*Argument)
-
+	ap := arg.(*Argument)
 	bat := proc.InputBatch()
 	if bat == nil {
+		if needToSendShuffledBats(ap) {
+			proc.SetInputBatches(ap.ctr.shuffledBats)
+		}
 		return true, nil
 	}
-
 	if bat.Length() == 0 {
 		bat.Clean(proc.Mp())
-		proc.SetInputBatch(batch.EmptyBatch)
+		proc.SetInputBatches([]*batch.Batch{batch.EmptyBatch})
 		return false, nil
 	}
-	proc.SetInputBatch(bat)
+	if ap.ShuffleType == int32(plan.ShuffleType_Hash) {
+		return hashShuffle(bat, ap, proc)
+	} else if ap.ShuffleType == int32(plan.ShuffleType_Range) {
+		return rangeShuffle(bat, ap, proc)
+	} else {
+		panic("unsupported shuffle type!")
+	}
+}
+
+func (arg *Argument) initShuffle() {
+	if arg.ctr.sels == nil {
+		arg.ctr.sels = make([][]int32, arg.AliveRegCnt)
+		for i := 0; i < int(arg.AliveRegCnt); i++ {
+			arg.ctr.sels[i] = make([]int32, 8192)
+		}
+		arg.ctr.batsCount = 0
+		arg.ctr.shuffledBats = make([]*batch.Batch, arg.AliveRegCnt)
+	}
+}
+
+func (arg *Argument) getSels() [][]int32 {
+	for i := range arg.ctr.sels {
+		arg.ctr.sels[i] = arg.ctr.sels[i][:0]
+	}
+	return arg.ctr.sels
+}
+
+func getShuffledSelsByHash(ap *Argument, bat *batch.Batch) [][]int32 {
+	sels := ap.getSels()
+	lenRegs := uint64(ap.AliveRegCnt)
+	groupByVec := bat.Vecs[ap.ShuffleColIdx]
+	switch groupByVec.GetType().Oid {
+	case types.T_int64:
+		groupByCol := vector.MustFixedCol[int64](groupByVec)
+		for row, v := range groupByCol {
+			regIndex := plan2.SimpleInt64HashToRange(uint64(v), lenRegs)
+			sels[regIndex] = append(sels[regIndex], int32(row))
+		}
+	case types.T_int32:
+		groupByCol := vector.MustFixedCol[int32](groupByVec)
+		for row, v := range groupByCol {
+			regIndex := plan2.SimpleInt64HashToRange(uint64(v), lenRegs)
+			sels[regIndex] = append(sels[regIndex], int32(row))
+		}
+	case types.T_int16:
+		groupByCol := vector.MustFixedCol[int16](groupByVec)
+		for row, v := range groupByCol {
+			regIndex := plan2.SimpleInt64HashToRange(uint64(v), lenRegs)
+			sels[regIndex] = append(sels[regIndex], int32(row))
+		}
+	case types.T_uint64:
+		groupByCol := vector.MustFixedCol[uint64](groupByVec)
+		for row, v := range groupByCol {
+			regIndex := plan2.SimpleInt64HashToRange(v, lenRegs)
+			sels[regIndex] = append(sels[regIndex], int32(row))
+		}
+	case types.T_uint32:
+		groupByCol := vector.MustFixedCol[uint32](groupByVec)
+		for row, v := range groupByCol {
+			regIndex := plan2.SimpleInt64HashToRange(uint64(v), lenRegs)
+			sels[regIndex] = append(sels[regIndex], int32(row))
+		}
+	case types.T_uint16:
+		groupByCol := vector.MustFixedCol[uint16](groupByVec)
+		for row, v := range groupByCol {
+			regIndex := plan2.SimpleInt64HashToRange(uint64(v), lenRegs)
+			sels[regIndex] = append(sels[regIndex], int32(row))
+		}
+	case types.T_char, types.T_varchar, types.T_text:
+		groupByCol := vector.MustFixedCol[types.Varlena](groupByVec)
+		for row, v := range groupByCol {
+			regIndex := plan2.SimpleCharHashToRange(v.GetByteSlice(groupByVec.GetArea()), lenRegs)
+			sels[regIndex] = append(sels[regIndex], int32(row))
+		}
+	default:
+		panic("unsupported shuffle type, wrong plan!") //something got wrong here!
+	}
+	return sels
+}
+
+func genShuffledBatsByHash(ap *Argument, bat *batch.Batch, proc *process.Process) error {
+	//release old bats
+	defer proc.PutBatch(bat)
+
+	lenVecs := len(bat.Vecs)
+	shuffledBats := ap.ctr.shuffledBats
+	if ap.ctr.batsCount == 0 {
+		//initialize shuffled bats
+		for regIndex := range shuffledBats {
+			shuffledBats[regIndex] = batch.NewWithSize(lenVecs)
+			shuffledBats[regIndex].ShuffleIDX = uint32(regIndex)
+			shuffledBats[regIndex].Zs = proc.Mp().GetSels()
+			for j := range shuffledBats[regIndex].Vecs {
+				shuffledBats[regIndex].Vecs[j] = proc.GetVector(*bat.Vecs[j].GetType())
+			}
+		}
+	}
+
+	sels := getShuffledSelsByHash(ap, bat)
+
+	//generate new shuffled bats
+	for regIndex := range shuffledBats {
+		lenSels := len(sels[regIndex])
+		if lenSels > 0 {
+			b := shuffledBats[regIndex]
+			for vecIndex := range b.Vecs {
+				v := b.Vecs[vecIndex]
+				err := v.Union(bat.Vecs[vecIndex], sels[regIndex], proc.Mp())
+				if err != nil {
+					return err
+				}
+			}
+			for i := 0; i < lenSels; i++ {
+				b.Zs = append(b.Zs, bat.Zs[sels[regIndex][i]])
+			}
+		}
+	}
+
+	ap.ctr.batsCount++
+	return nil
+}
+
+// accumulate enough batch size
+func needToSendShuffledBats(ap *Argument) bool {
+	if ap.ctr.batsCount > 0 {
+		maxSize := 0
+		for i := range ap.ctr.shuffledBats {
+			if ap.ctr.shuffledBats[i].Length() > maxSize {
+				maxSize = ap.ctr.shuffledBats[i].Length()
+			}
+		}
+		if maxSize > shuffleBatchSize {
+			return true
+		}
+	}
+	return false
+}
+
+func hashShuffle(bat *batch.Batch, ap *Argument, proc *process.Process) (bool, error) {
+	err := genShuffledBatsByHash(ap, bat, proc)
+	if err != nil {
+		return false, err
+	}
+	if needToSendShuffledBats(ap) {
+		proc.SetInputBatches(ap.ctr.shuffledBats)
+	} else {
+		proc.SetInputBatches([]*batch.Batch{batch.EmptyBatch})
+	}
+	return false, nil
+}
+
+func allBatchInOneRange(ap *Argument, bat *batch.Batch) (bool, uint64) {
+	lenRegs := uint64(ap.AliveRegCnt)
+	groupByVec := bat.Vecs[ap.ShuffleColIdx]
+	var regIndexFirst, regIndexLast uint64
+	switch groupByVec.GetType().Oid {
+	case types.T_int64:
+		groupByCol := vector.MustFixedCol[int64](groupByVec)
+		vfirst := groupByCol[0]
+		vlast := groupByCol[groupByVec.Length()-1]
+		regIndexFirst = plan2.GetRangeShuffleIndexSigned(ap.ShuffleColMin, ap.ShuffleColMax, vfirst, lenRegs)
+		regIndexLast = plan2.GetRangeShuffleIndexSigned(ap.ShuffleColMin, ap.ShuffleColMax, vlast, lenRegs)
+
+	case types.T_int32:
+		groupByCol := vector.MustFixedCol[int32](groupByVec)
+		vfirst := groupByCol[0]
+		vlast := groupByCol[groupByVec.Length()-1]
+		regIndexFirst = plan2.GetRangeShuffleIndexSigned(ap.ShuffleColMin, ap.ShuffleColMax, int64(vfirst), lenRegs)
+		regIndexLast = plan2.GetRangeShuffleIndexSigned(ap.ShuffleColMin, ap.ShuffleColMax, int64(vlast), lenRegs)
+	case types.T_int16:
+		groupByCol := vector.MustFixedCol[int16](groupByVec)
+		vfirst := groupByCol[0]
+		vlast := groupByCol[groupByVec.Length()-1]
+		regIndexFirst = plan2.GetRangeShuffleIndexSigned(ap.ShuffleColMin, ap.ShuffleColMax, int64(vfirst), lenRegs)
+		regIndexLast = plan2.GetRangeShuffleIndexSigned(ap.ShuffleColMin, ap.ShuffleColMax, int64(vlast), lenRegs)
+	case types.T_uint64:
+		groupByCol := vector.MustFixedCol[uint64](groupByVec)
+		vfirst := groupByCol[0]
+		vlast := groupByCol[groupByVec.Length()-1]
+		regIndexFirst = plan2.GetRangeShuffleIndexUnsigned(uint64(ap.ShuffleColMin), uint64(ap.ShuffleColMax), vfirst, lenRegs)
+		regIndexLast = plan2.GetRangeShuffleIndexUnsigned(uint64(ap.ShuffleColMin), uint64(ap.ShuffleColMax), vlast, lenRegs)
+	case types.T_uint32:
+		groupByCol := vector.MustFixedCol[uint32](groupByVec)
+		vfirst := groupByCol[0]
+		vlast := groupByCol[groupByVec.Length()-1]
+		regIndexFirst = plan2.GetRangeShuffleIndexUnsigned(uint64(ap.ShuffleColMin), uint64(ap.ShuffleColMax), uint64(vfirst), lenRegs)
+		regIndexLast = plan2.GetRangeShuffleIndexUnsigned(uint64(ap.ShuffleColMin), uint64(ap.ShuffleColMax), uint64(vlast), lenRegs)
+	case types.T_uint16:
+		groupByCol := vector.MustFixedCol[uint16](groupByVec)
+		vfirst := groupByCol[0]
+		vlast := groupByCol[groupByVec.Length()-1]
+		regIndexFirst = plan2.GetRangeShuffleIndexUnsigned(uint64(ap.ShuffleColMin), uint64(ap.ShuffleColMax), uint64(vfirst), lenRegs)
+		regIndexLast = plan2.GetRangeShuffleIndexUnsigned(uint64(ap.ShuffleColMin), uint64(ap.ShuffleColMax), uint64(vlast), lenRegs)
+	default:
+		panic("unsupported shuffle type, wrong plan!") //something got wrong here!
+	}
+	if regIndexFirst == regIndexLast {
+		return true, regIndexFirst
+	} else {
+		return false, 0
+	}
+}
+
+func getShuffledSelsByRange(ap *Argument, bat *batch.Batch) [][]int32 {
+	sels := ap.getSels()
+	lenRegs := uint64(ap.AliveRegCnt)
+	groupByVec := bat.Vecs[ap.ShuffleColIdx]
+	switch groupByVec.GetType().Oid {
+	case types.T_int64:
+		groupByCol := vector.MustFixedCol[int64](groupByVec)
+		for row, v := range groupByCol {
+			regIndex := plan2.GetRangeShuffleIndexSigned(ap.ShuffleColMin, ap.ShuffleColMax, v, lenRegs)
+			sels[regIndex] = append(sels[regIndex], int32(row))
+		}
+	case types.T_int32:
+		groupByCol := vector.MustFixedCol[int32](groupByVec)
+		for row, v := range groupByCol {
+			regIndex := plan2.GetRangeShuffleIndexSigned(ap.ShuffleColMin, ap.ShuffleColMax, int64(v), lenRegs)
+			sels[regIndex] = append(sels[regIndex], int32(row))
+		}
+	case types.T_int16:
+		groupByCol := vector.MustFixedCol[int16](groupByVec)
+		for row, v := range groupByCol {
+			regIndex := plan2.GetRangeShuffleIndexSigned(ap.ShuffleColMin, ap.ShuffleColMax, int64(v), lenRegs)
+			sels[regIndex] = append(sels[regIndex], int32(row))
+		}
+	case types.T_uint64:
+		groupByCol := vector.MustFixedCol[uint64](groupByVec)
+		for row, v := range groupByCol {
+			regIndex := plan2.GetRangeShuffleIndexUnsigned(uint64(ap.ShuffleColMin), uint64(ap.ShuffleColMax), v, lenRegs)
+			sels[regIndex] = append(sels[regIndex], int32(row))
+		}
+	case types.T_uint32:
+		groupByCol := vector.MustFixedCol[uint32](groupByVec)
+		for row, v := range groupByCol {
+			regIndex := plan2.GetRangeShuffleIndexUnsigned(uint64(ap.ShuffleColMin), uint64(ap.ShuffleColMax), uint64(v), lenRegs)
+			sels[regIndex] = append(sels[regIndex], int32(row))
+		}
+	case types.T_uint16:
+		groupByCol := vector.MustFixedCol[uint16](groupByVec)
+		for row, v := range groupByCol {
+			regIndex := plan2.GetRangeShuffleIndexUnsigned(uint64(ap.ShuffleColMin), uint64(ap.ShuffleColMax), uint64(v), lenRegs)
+			sels[regIndex] = append(sels[regIndex], int32(row))
+		}
+	default:
+		panic("unsupported shuffle type, wrong plan!") //something got wrong here!
+	}
+	return sels
+}
+
+func genShuffledBatsByRange(ap *Argument, bat *batch.Batch, proc *process.Process) error {
+	//release old bats
+	defer proc.PutBatch(bat)
+
+	lenVecs := len(bat.Vecs)
+	shuffledBats := ap.ctr.shuffledBats
+	if ap.ctr.batsCount == 0 {
+		//initialize shuffled bats
+		for regIndex := range shuffledBats {
+			shuffledBats[regIndex] = batch.NewWithSize(lenVecs)
+			shuffledBats[regIndex].ShuffleIDX = uint32(regIndex)
+			shuffledBats[regIndex].Zs = proc.Mp().GetSels()
+			for j := range shuffledBats[regIndex].Vecs {
+				shuffledBats[regIndex].Vecs[j] = proc.GetVector(*bat.Vecs[j].GetType())
+			}
+		}
+	}
+
+	sels := getShuffledSelsByRange(ap, bat)
+
+	//generate new shuffled bats
+	for regIndex := range shuffledBats {
+		lenSels := len(sels[regIndex])
+		if lenSels > 0 {
+			b := shuffledBats[regIndex]
+			for vecIndex := range b.Vecs {
+				v := b.Vecs[vecIndex]
+				err := v.Union(bat.Vecs[vecIndex], sels[regIndex], proc.Mp())
+				if err != nil {
+					return err
+				}
+			}
+			for i := 0; i < lenSels; i++ {
+				b.Zs = append(b.Zs, bat.Zs[sels[regIndex][i]])
+			}
+		}
+	}
+
+	ap.ctr.batsCount++
+	return nil
+}
+
+func rangeShuffle(bat *batch.Batch, ap *Argument, proc *process.Process) (bool, error) {
+	groupByVec := bat.Vecs[ap.ShuffleColIdx]
+	if groupByVec.GetSorted() {
+		ok, regIndex := allBatchInOneRange(ap, bat)
+		if ok {
+			bat.ShuffleIDX = uint32(regIndex)
+			proc.SetInputBatches([]*batch.Batch{bat})
+			return false, nil
+		}
+	}
+
+	err := genShuffledBatsByRange(ap, bat, proc)
+	if err != nil {
+		return false, err
+	}
+	if needToSendShuffledBats(ap) {
+		proc.SetInputBatches(ap.ctr.shuffledBats)
+	} else {
+		proc.SetInputBatches([]*batch.Batch{batch.EmptyBatch})
+	}
 	return false, nil
 }
